+++ internal/bills/files_test.go
@@ -1,4 +1,4 @@
-package gha
+package bills
 
 import (
 	"os"
